chore : refinement + fixes
