+++ src/app/(landing)/blog/all/layout.module.scss
@@ -1,0 +1,4 @@
+.element {
+  background-color: #ece3f6;
+  color: var(--Black, #000);
+}
