--- src/app/(landing)/blog/page.tsx
@@ -1,5 +1,0 @@
-import LatestBlogPosts from './components/latest-blog-posts';
-
-export default function BlogLandingPage() {
-  return <LatestBlogPosts />;
-}
